renamed apiEnpoints.jsx to endpoints.js
